util: refactor 7702 authorization lists to util package
