+++ packages/tx/src/types.ts
@@ -4,6 +4,8 @@ import type { Common, Hardfork, ParamsDict } from '@ethereumjs/common'
 import type {
   Address,
   AddressLike,
+  AuthorizationList,
+  AuthorizationListBytes,
   BigIntLike,
   BytesLike,
   PrefixedHexString,
@@ -605,32 +607,3 @@ export type AccessListItem = {
 export type AccessListBytesItem = [Uint8Array, Uint8Array[]]
 export type AccessListBytes = AccessListBytesItem[]
 export type AccessList = AccessListItem[]
-
-/**
- * Authorization list types
- */
-export type AuthorizationListItemUnsigned = {
-  chainId: PrefixedHexString
-  address: PrefixedHexString
-  nonce: PrefixedHexString
-}
-
-export type AuthorizationListItem = {
-  yParity: PrefixedHexString
-  r: PrefixedHexString
-  s: PrefixedHexString
-} & AuthorizationListItemUnsigned
-
-// Tuple of [chain_id, address, [nonce], y_parity, r, s]
-export type AuthorizationListBytesItem = [
-  Uint8Array,
-  Uint8Array,
-  Uint8Array,
-  Uint8Array,
-  Uint8Array,
-  Uint8Array,
-]
-export type AuthorizationListBytes = AuthorizationListBytesItem[]
-export type AuthorizationList = AuthorizationListItem[]
-
-export type AuthorizationListBytesItemUnsigned = [Uint8Array, Uint8Array, Uint8Array]
