+++ packages/tx/src/util/index.ts
@@ -1,4 +1,3 @@
 // Do not add `./internal.ts`, this will export the internal helpers also at package level
 export * from './general.ts'
 export * from './access.ts'
-export * from './authorization.ts'
